+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added support for loading `libvpx` from a custom path using the `panvpx.libvpx.path` system property.
+
 ## [0.3.6] — 2026-05-28
 
 ### Changed